328. Odd Even Linked List
Given a singly linked list, group all odd nodes together followed by the
even nodes. Please note here we are talking about the node number and
not the value in the nodes.

You should try to do it in place. The program should run in O(1) space
complexity and O(nodes) time complexity.